Move SocketFactory to Jellyfin.Networking
